Refactor amount input field to use basic input field
