+++ src/components/ExpenseForm/ParticipantsSection/ParticipantsSection.jsx
@@ -1,4 +1,3 @@
-import { useState } from 'react';
 import { Divider, IconButton } from '@mui/material';
 import Grid from '@mui/material/Grid2';
 import AddCircleIcon from '@mui/icons-material/AddCircle';
